requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
keyboard==0.13.5
MarkupSafe==2.1.5
mss==9.0.1
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pynput==1.7.7
pyobjc==10.3
pyobjc-core==10.3
pyobjc-framework-Accessibility==10.3
pyobjc-framework-Accounts==10.3
pyobjc-framework-AddressBook==10.3
pyobjc-framework-AdServices==10.3
pyobjc-framework-AdSupport==10.3
pyobjc-framework-AppleScriptKit==10.3
pyobjc-framework-AppleScriptObjC==10.3
pyobjc-framework-ApplicationServices==10.3
pyobjc-framework-AppTrackingTransparency==10.3
pyobjc-framework-AudioVideoBridging==10.3
pyobjc-framework-AuthenticationServices==10.3
pyobjc-framework-AutomaticAssessmentConfiguration==10.3
pyobjc-framework-Automator==10.3
pyobjc-framework-AVFoundation==10.3
pyobjc-framework-AVKit==10.3
pyobjc-framework-AVRouting==10.3
pyobjc-framework-BackgroundAssets==10.3
pyobjc-framework-BusinessChat==10.3
pyobjc-framework-CalendarStore==10.3
pyobjc-framework-CallKit==10.3
pyobjc-framework-CFNetwork==10.3
pyobjc-framework-ClassKit==10.3
pyobjc-framework-CloudKit==10.3
pyobjc-framework-Cocoa==10.3
pyobjc-framework-Collaboration==10.3
pyobjc-framework-ColorSync==10.3
pyobjc-framework-Contacts==10.3
pyobjc-framework-ContactsUI==10.3
pyobjc-framework-CoreAudio==10.3
pyobjc-framework-CoreAudioKit==10.3
pyobjc-framework-CoreBluetooth==10.3
pyobjc-framework-CoreData==10.3
pyobjc-framework-CoreHaptics==10.3
pyobjc-framework-CoreLocation==10.3
pyobjc-framework-CoreMedia==10.3
pyobjc-framework-CoreMediaIO==10.3
pyobjc-framework-CoreMIDI==10.3
pyobjc-framework-CoreML==10.3
pyobjc-framework-CoreMotion==10.3
pyobjc-framework-CoreServices==10.3
pyobjc-framework-CoreSpotlight==10.3
pyobjc-framework-CoreText==10.3
pyobjc-framework-CoreWLAN==10.3
pyobjc-framework-CryptoTokenKit==10.3
pyobjc-framework-DataDetection==10.3
pyobjc-framework-DeviceCheck==10.3
pyobjc-framework-DictionaryServices==10.3
pyobjc-framework-DiscRecording==10.3
pyobjc-framework-DiscRecordingUI==10.3
pyobjc-framework-DiskArbitration==10.3
pyobjc-framework-DVDPlayback==10.3
pyobjc-framework-EventKit==10.3
pyobjc-framework-ExceptionHandling==10.3
pyobjc-framework-ExecutionPolicy==10.3
pyobjc-framework-ExtensionKit==10.3
pyobjc-framework-ExternalAccessory==10.3
pyobjc-framework-FileProvider==10.3
pyobjc-framework-FileProviderUI==10.3
pyobjc-framework-FinderSync==10.3
pyobjc-framework-FSEvents==10.3
pyobjc-framework-GameCenter==10.3
pyobjc-framework-GameController==10.3
pyobjc-framework-GameKit==10.3
pyobjc-framework-GameplayKit==10.3
pyobjc-framework-HealthKit==10.3
pyobjc-framework-ImageCaptureCore==10.3
pyobjc-framework-InputMethodKit==10.3
pyobjc-framework-InstallerPlugins==10.3
pyobjc-framework-InstantMessage==10.3
pyobjc-framework-Intents==10.3
pyobjc-framework-IntentsUI==10.3
pyobjc-framework-IOBluetooth==10.3
pyobjc-framework-IOBluetoothUI==10.3
pyobjc-framework-IOSurface==10.3
pyobjc-framework-iTunesLibrary==10.3
pyobjc-framework-KernelManagement==10.3
pyobjc-framework-LatentSemanticMapping==10.3
pyobjc-framework-LaunchServices==10.3
pyobjc-framework-libdispatch==10.3
pyobjc-framework-libxpc==10.3
pyobjc-framework-LinkPresentation==10.3
pyobjc-framework-LocalAuthentication==10.3
pyobjc-framework-LocalAuthenticationEmbeddedUI==10.3
pyobjc-framework-MailKit==10.3
pyobjc-framework-MapKit==10.3
pyobjc-framework-MediaAccessibility==10.3
pyobjc-framework-MediaLibrary==10.3
pyobjc-framework-MediaPlayer==10.3
pyobjc-framework-MediaToolbox==10.3
pyobjc-framework-Metal==10.3
pyobjc-framework-MetalFX==10.3
pyobjc-framework-MetalKit==10.3
pyobjc-framework-MetalPerformanceShaders==10.3
pyobjc-framework-MetalPerformanceShadersGraph==10.3
pyobjc-framework-MetricKit==10.3
pyobjc-framework-MLCompute==10.3
pyobjc-framework-ModelIO==10.3
pyobjc-framework-MultipeerConnectivity==10.3
pyobjc-framework-NaturalLanguage==10.3
pyobjc-framework-NetFS==10.3
pyobjc-framework-Network==10.3
pyobjc-framework-NetworkExtension==10.3
pyobjc-framework-NotificationCenter==10.3
pyobjc-framework-OpenDirectory==10.3
pyobjc-framework-OSAKit==10.3
pyobjc-framework-OSLog==10.3
pyobjc-framework-PassKit==10.3
pyobjc-framework-PencilKit==10.3
pyobjc-framework-PHASE==10.3
pyobjc-framework-Photos==10.3
pyobjc-framework-PhotosUI==10.3
pyobjc-framework-PreferencePanes==10.3
pyobjc-framework-PushKit==10.3
pyobjc-framework-Quartz==10.3
pyobjc-framework-QuickLookThumbnailing==10.3
pyobjc-framework-ReplayKit==10.3
pyobjc-framework-SafariServices==10.3
pyobjc-framework-SafetyKit==10.3
pyobjc-framework-SceneKit==10.3
pyobjc-framework-ScreenCaptureKit==10.3
pyobjc-framework-ScreenSaver==10.3
pyobjc-framework-ScreenTime==10.3
pyobjc-framework-ScriptingBridge==10.3
pyobjc-framework-SearchKit==10.3
pyobjc-framework-Security==10.3
pyobjc-framework-SecurityFoundation==10.3
pyobjc-framework-SecurityInterface==10.3
pyobjc-framework-ServiceManagement==10.3
pyobjc-framework-SharedWithYou==10.3
pyobjc-framework-SharedWithYouCore==10.3
pyobjc-framework-ShazamKit==10.3
pyobjc-framework-Social==10.3
pyobjc-framework-SoundAnalysis==10.3
pyobjc-framework-Speech==10.3
pyobjc-framework-SpriteKit==10.3
pyobjc-framework-StoreKit==10.3
pyobjc-framework-SyncServices==10.3
pyobjc-framework-SystemConfiguration==10.3
pyobjc-framework-SystemExtensions==10.3
pyobjc-framework-ThreadNetwork==10.3
pyobjc-framework-UniformTypeIdentifiers==10.3
pyobjc-framework-UserNotifications==10.3
pyobjc-framework-UserNotificationsUI==10.3
pyobjc-framework-VideoSubscriberAccount==10.3
pyobjc-framework-VideoToolbox==10.3
pyobjc-framework-Virtualization==10.3
pyobjc-framework-Vision==10.3
pyobjc-framework-WebKit==10.3
pytesseract==0.3.10
six==1.16.0
Werkzeug==3.0.3
